MYR-529 Move the wake step to its own file to hold the 300-line cap
Review warning: setup_completion.go crossed 300 lines at 304. The wake is a
self-contained step of the sequence and splits cleanly, the same way configure
already did.
