+++ internal/telemetry/setup_completion.go
@@ -207,33 +207,6 @@ func (s *SetupCompleter) Complete(ctx context.Context, row VehicleSnapshotRow) (
 	return st, err
 }
 
-// wake issues the unsigned wake and reports whether Tesla ACCEPTED it.
-//
-// The returned vehicle state is logged and otherwise ignored: for a sleeping
-// car Tesla answers with the pre-wake state, so gating on `online` would refuse
-// every car this endpoint exists to serve (see fleet_api_vehicle_wake.go).
-func (s *SetupCompleter) wake(ctx context.Context, token string, row VehicleSnapshotRow, vin string) bool {
-	callCtx, cancel := context.WithTimeout(ctx, s.cfg.CallTimeout)
-	defer cancel()
-
-	state, err := s.deps.Probe.WakeVehicle(callCtx, token, row.VIN)
-	if err != nil {
-		s.logger.Warn("complete-setup: wake was refused",
-			slog.String("vehicle_id", row.ID), slog.String("vin", vin),
-			slog.String("error", redactedErrorText(err)))
-		return false
-	}
-	reported := ""
-	if state != nil {
-		reported = state.State
-	}
-	s.logger.Info("complete-setup: wake accepted",
-		slog.String("event", "setup_complete_wake_accepted"),
-		slog.String("vehicle_id", row.ID), slog.String("vin", vin),
-		slog.String("reported_state", reported))
-	return true
-}
-
 // probePairing polls `fleet_status` until Tesla reports the key enrolled or the
 // budget expires.
 //
